% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_quarto_ext.R
\name{available_extensions}
\alias{available_extensions}
\title{List bundled Quarto extensions}
\usage{
available_extensions()
}
\value{
a string of available extensions for install
}
\description{
List bundled Quarto extensions
}
