% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_quarto_ext.R
\name{use_quarto_ext}
\alias{use_quarto_ext}
\title{Install bundled Quarto extensions into current working directory and create
new qmd using skeleton documents. This function extends a function written by
Thomas Mock: https://github.com/jthomasmock/octavo/blob/master/R/use_quarto_ext.R}
\usage{
use_quarto_ext(file_name = NULL, ext_name = "quartotemplate")
}
\arguments{
\item{file_name}{Name of new qmd file and sub-directory to be created}

\item{ext_name}{String indicating which extension to install}
}
\value{
a message if extension was successfully copied over
}
\description{
Install bundled Quarto extensions into current working directory and create
new qmd using skeleton documents. This function extends a function written by
Thomas Mock: https://github.com/jthomasmock/octavo/blob/master/R/use_quarto_ext.R
}
